Add files via upload
The project consists of 5 separate files that work together to be ran and compiled under the Game.java file which is included within those 5 files. All of the battle sequences can be thought of to be contained and called from the Game file whereas the attacks and damage points of those attacks along with their HP and chances of a critical hit are set in the files Queen and Alice. The Worldmap is what ties together all of the events and text in the game to all of the other files, and contains the most content which therefor may be perceived as the most important piece of the puzzle. The inventory file is just setting and the items and their attributes which is just +HP or –HP, and are called into the Worldmap file in proper sequence to the story.
The importance of getters and setters is very high in this. The world map file contains all of the text outputted to the command prompt, or in other words the story. What it also does is use getters and setters to pair the output with the correct sequences to advance you forth in the game. The return function at the end of the methods statements shows take guide the character to either the next level or alternate routes. In a sense the return function is like saying “go to”. These functions are within while loops and if statements, but also are also at the end of all levels even if there is no user input. TextIO.getln will also prompt the user for their input, with the if statements setup in a while loop paired with this function the if statements are the options the user has to choose from, which will be seen in the “TextIO.putln” which is really the opposite feature to output prewritten text for the user to see or interact with.
One new addition that’s different from the previous version of this game is the delay in writing being output. I think this adds a nice touch because it allows your eyes to fully read what the options are or text is describing before the next level. This took some time to get right as the delay was too long at first and made it seem like an unnecessarily long pause after every few lines or command. The player will think the game has frozen and start hitting buttons to get it to unfreeze but then likely end up making an invalid entry and starting an endless loop of frustration. It delays 3 seconds for every input and during the attack sequences each attack is output in .5 seconds. The try{Thread.sleep(3000);} controls the initial delay whereas the battle delay is controlled by try{Thread.sleep(500);} within the while fight loop. The closing blocks are written directly after the semi colon which looks strange but it is the same as putting the block on the next line and leaving the semi colon.
When compiling the Game, testing and debugging in the command prompt shows us exactly what file and the line in that file where the error occurs. The number of errors in this file were largely due to the immense number of constructors used just being copy and pasted after the first written one, as the setup is the same for each. I would do this then rewrite the pasted lines to input the rest of the data. This had created major issues during the parsing of the compiler, which was solved eventually by looking at the left-hand side where the lines are collapsible starting from the blocks implemented.  I had missed so many blocks and couldn't make heads or tails of how many errors I was seeing until I tried this trick. This was an easier way to see if the blocks lined up perfectly because with this project there were around 40 errors upon the first attempt at compiling. The collapsible lines could separate any statements or main classes and show clearly if the statements bled into another accidentally by the way the lined up when minimized.  	
Ultimately the idea was to allow the player a few different options while making the trajectory forward easy at first glance, but their story will come to an end quick if that is all they do. The rudimentary design of a multitude of paths for the player to take displays the fundamentals in programming for this assignment. The idea is that if you don’t pick up any of several items which increase either your attack power or hit points then you die very quick, and even if you get passed the first boss the player is still going to need to pick up more items after that point. Alice has a random100 method, it will pull a random number which is associate with either strong attack, weak attack or miss. Each boss gets progressively harder by design, I felt like this is what really made it a game rather than just a story. The implementation of the automated fighting style eliminates an element of user input which effects the game, but this being taken out still has a challenge in the character choosing the right paths and likely having to retry the game multiple times before finally defeating the Queen. The items all add attack power each one you get such as the chainsaw, dagger or increase HP like with food or potions, and there is no ultimate limit on either so they will just keep getting higher the more you get making the battles quicker and taking less % damage. The attack multiplier is controlled by if statements within the Game file that work like saying if Alice gets the rope *2 to get.Attack, and it does this for every weapon, even the rope and lighter. Its implying that you can use your imagination to envision what kind of maneuver would be executed with both of these objects separately or combined, although the other weapons are straight forward. The rope here does have dual functions which is the only weapon to feature this, but it is used to scale the cliff leading to the castle and also used during any fight as an added attack bonus.